주문시 고객 정보도 수정되도록 수정, 코드 실수 수정, 고객 데이터 시딩 간략화, 주문 가격 타입 변경

+++ src/main/java/e3i2/ecommerce_backoffice/domain/customer/dto/GetCustomerResponse.java
@@ -17,9 +17,9 @@ public class GetCustomerResponse {
     @JsonFormat(shape = JsonFormat.Shape.STRING, pattern = "yyyy-MM-dd")
     private final LocalDateTime createdAt;
     private final Long totalOrders;
-    private final BigInteger totalSpent;
+    private final Long totalSpent;
 
-    public GetCustomerResponse(Long customerId, String customerName, String email, String phone, String status, LocalDateTime createdAt, Long totalOrders, BigInteger totalSpent) {
+    public GetCustomerResponse(Long customerId, String customerName, String email, String phone, String status, LocalDateTime createdAt, Long totalOrders, Long totalSpent) {
         this.customerId = customerId;
         this.customerName = customerName;
         this.email = email;
